vaults move: ../data_sources/[dsId]/documents/[documentId]/index
